requirements.txt
# Data collection
google-api-python-client>=2.0.0
praw>=7.7.0

# NLP / ABSA
sentence-transformers>=2.2.0
vaderSentiment>=3.3.2
scikit-learn>=1.3.0
nltk>=3.8.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.66.0
python-dotenv>=1.0.0

# Backend (Phase 2)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
anthropic>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
